New test related to stackoverflow question:
In PDFBox, how to create a link annotation with “rollover” / “mouse
over” effects?
http://stackoverflow.com/questions/28144941/in-pdfbox-how-to-create-a-link-annotation-with-rollover-mouse-over-effect

testCreateRolloverAnnotation() attempts to create a PDF with the desired
rollover effect using a PDAnnotationLink. Adobe viewers don't show the
effect.

testCreateRolloverButton() attempts the same using a PDAnnotationWidget
customized as a pushbutton. Adobe viewers do show the effect.