Added test SignLikeLoneWolf focusing on the stackoverflow question
Signing PDF with PDFBox and BouncyCastle
http://stackoverflow.com/questions/30400728/signing-pdf-with-pdfbox-and-bouncycastle

This class tests the OP's original code (CreateSignature with
CreateSignature.fixed = false) and the fixed code (CreateSignature with
CreateSignature.fixed = true).

The issue essentially is that PDDocument.saveIncremental(InputStream,
OutputStream) expects the InputStream to cover the PDF including
preparations for saving while the OP only provided an InputStream
covering the original file.