New tests SignLikeUnOriginalToo.* focusing on the stackoverflow question
Attachment damages signature part 2
http://stackoverflow.com/questions/30549830/attachment-damages-signature-part-2

The OP's original code is in TC3. From there the method TC3.doSign() is
copied here with different degrees of manipulations,doSignOriginal being
essentially the original code with dynamic input streams and output
file, doSignRemoveType fixes the misleading trailer Type entry resulting
in valid signatures, doSignTwoRevisions tries to use an incremental
update for the first change, too, but fails, and doSignOneStep tries to
add the image in the same incremental update as the image but fails for
cross reference table sources.