New test ClipPage.testTestTop focusing on the stackoverflow question
Splitting at a specific point in PDFBox
http://stackoverflow.com/questions/30616220/splitting-at-a-specific-point-in-pdfbox

One way to achieve the task, i.e. to split a page at a certain point
(i.e. all content above a limit to be included and everything below to
be excluded) would be to prepend a clip path. This is implemented in
clipPage(PDDocument, PDPage, BoundingBox).