New test SignLikeDenisMP focusing on the stackoverflow question
How can I get PDVisibleSigProperties to write the signature on the 3
page into the signature box
http://stackoverflow.com/questions/31571055/how-can-i-get-pdvisiblesigproperties-to-write-the-signature-on-the-3-page-into-t
hhs-745.pdf
http://www.ors.od.nih.gov/ser/dpsac/bgchecks/Documents/hhs-745.pdf

The method testOriginalA() essentially represents the code of the
author's first code version

Running it that method indeed showed that there are some issues in the
PDFBox visible signature code. Especially it seems to essentially ignore
the signature field name set using
<code>.signatureFieldName("ApplicantSignature")</code>. Internally it
looks like that name is used once to set the signature dictionary "Name"
value but even that seems to get lost.

Beware, this test requires older BouncyCastle versions to run,
officially 1.44 but 1.47 seems to work, too.