New test IncrementalFormFill focusing on the stackoverflow question
PDFBox 1.8.10: Fill and Sign Document, Filling again fails
http://stackoverflow.com/questions/33011970/pdfbox-1-8-10-fill-and-sign-document-filling-again-fails
doc_v2.pdf
https://www.dropbox.com/s/xf5pb0ng8k9zd4i/doc_v2.pdf?dl=0
doc_v2_fillsigned.pdf
https://www.dropbox.com/s/s8295tfyjpe1l4l/doc_v2_fillsigned.pdf?dl=0

The cause of the problem is that somehow during the original filling and
signing the fonts in the default resources of the interactive form
dictionary got lost.

In the original document doc_v2.pdf the interactive form dictionary
contains entries for ZaDb and Helv in the Font dictionary in the default
resources DR dictionary; so form filli-in works, cf.
testFillWithoutFillAndSign.

In contrast in the interactive form dictionary of the filled and signed
document doc_v2_fillsigned.pdf the Font dictionary in the default
resources DR dictionary is missing; so form fill-in fails, cf.
testFillAfterFillAndSign.