New test SignLikePdfboxExample focusing on the stackoverflow question
PDFBox 1.8.10: Fill and Sign Document, Filling again fails
http://stackoverflow.com/questions/33011970/pdfbox-1-8-10-fill-and-sign-document-filling-again-fails
doc_v2.pdf
https://www.dropbox.com/s/xf5pb0ng8k9zd4i/doc_v2.pdf?dl=0

The cause of the problem is that somehow during the original filling and
signing the fonts in the default resources of the interactive form
dictionary got lost.

PDFBox while filling in the form tries to access the font definition to
create an appearance stream. It doesn't find it and, therefore,
eventually fails.

I simply applied the PDFBox example CreateVisibleSignature to the OP's
original doc_v2.pdf file. Indeed, this already removed the Font default
resource dictionary.

Beware, this test requires older BouncyCastle versions to run,
officially 1.44 but 1.47 seems to work, too.