New ExtractText test testExtractStyledFromExampleDocument focusing on the stackoverflow question
Detect Bold, Italic and Strike Through text using PDFBox with VB.NET
http://stackoverflow.com/questions/39962563/detect-bold-italic-and-strike-through-text-using-pdfbox-with-vb-net
Example Document.pdf
http://www.filedropper.com/exampledocument

This test shows how to extract text with styles using the
PDFStyledTextStripper.

This PDFTextStripper subclass extracts text with the styles used in the
example document.

BEWARE: This merely is a proof of concept! In particular the tests in TransformedRectangle.underlines(TextPosition) and TransformedRectangle.strikesThrough(TextPosition) only work with the restriction described in the methods themselves.