New AddContentToOCG test testAddContentToNewOrExistingOCG focusing on the stack overflow question
How do I make modifications to existing layer(Optional Content Group) in pdf?
http://stackoverflow.com/questions/43275212/how-do-i-make-modifications-to-existing-layeroptional-content-group-in-pdf

This test show how add content to an optional content group which may or may not already exist. The main intelligence is in the helper method addTextToLayer.


As an aside updated the PDFBox version tested against to 1.8.12.