New CreateCheckbox test testCheckboxLikeSureshGoud focusing on the stack overflow question
Creating a checkbox and printing it to pdf file is not working using pdfbox 1.8.9 api
http://stackoverflow.com/questions/43604973/creating-a-checkbox-and-printing-it-to-pdf-file-is-not-working-using-pdfbox-1-8

This test executes the OP's method for checkbox creation.

As it turns out it has some errors: It treats a checkbox field as a text field for which a PDF reader should create an appearance based on the default appearance (DA) value of the field in particular if NeedAppearances is true. Checkboxes are different, though: one does have to supply an appearance stream at least for the on state.