New OpenFile test focusing on the stack overflow question
PDF failing to be loaded because it can't be read by PDDocument
http://stackoverflow.com/questions/44042244/pdf-failing-to-be-loaded-because-it-cant-be-read-by-pddocument
CG_PARTANCE_701018_0415.pdf
https://drive.google.com/open?id=0B57cf1nqGbC4SEYtemxwQjFQNkU

The PDF could properly be loaded. The code then choked on acroForm.getFields() which threw a NullPointerException, and it did so for the simple reason that the PDF contains no form definition and, therefore, acroForm was null.